+++ src/Services/Job/JobCompany/JobCompany.Business/Dtos/ExamDtos/GetExamQuestionsDetailDto.cs
@@ -6,6 +6,6 @@ public class GetExamQuestionsDetailDto
 {
     public int TotalQuestions { get; set; }
     public decimal LimitRate { get; set; }
-    public byte? Duration { get; set; }
+    public short? Duration { get; set; }
     public List<QuestionPublicDto> Questions { get; set; }
 }
